requirements.txt
autogen-agentchat>=0.5.0
autogen-ext>=0.5.0
python-dotenv>=1.0.0 